specs: ignore celebration titles when comparing latest state
it would be desirable to have the Latin celebration titles matching
between calendarium-romanum and calendarium-romanum-historical,
but for now that's not an issue to be addressed soon.

Note: spec/data_spec.rb pass with fresh development version of
calendarium-romanum, but not with the last version currently released
(0.9.0) which lacks a few data fixes.